Add modal to utilization page to edit silo quotas
